Fixing ITrackingAware persistance
When persisiting an object, the Tracker class was looking for the
configuration for the target type, but when using ITrackingAware,
it's the target instance that owns the configuration, not the target's
type. This commit fixes that bug in Tracker.cs and adds a unit test to
cover ITrackingAware.

Changing version number to 2.1.7.