TrackingConfigurationGeneric properly implemented
The generic version of TrackingConfiguration was supposed to be a
decorator around the non-generic one. However the implementation was
both messy and broken - it was deriving from TrackingConfiguration as
well as wrapping it. The problem was it was not delegating all methods
to the wrapped instance. For example, when it was forwarding Id to the
wrapped class, but was not forwarding GetStoreId to the wrapped class.

This commit implements the generic version as a proper decorator. A new
interface has been extracted from the TrackingConfiguration and this
interface is implemented in both TrackingConfiguration and
TrackingConfiguration<> classes. The generic version now implements all
members of this interface and delegates all of them to the wrapped class
while exposing additional (typed) configuration methods.