Fix: type conversions when setting property value
The bug this commit solves is described here:
https://github.com/anakic/Jot/issues/58

It was caused by the fact that Convert.ChangeType fails if the object is
of a derived type (derived from the property type) but in this situation
no conversion should occur. Additionally, the value conversion logic
should occur for properties that have [DefaultValue] specified as well,
but in the previous version this was not done.